+++ admin/srai_lookup.php
@@ -2,7 +2,7 @@
 /***************************************
  * http://www.program-o.com
  * PROGRAM O
- * Version: 2.6.3
+ * Version: 2.6.5
  * FILE: srai_lookup.php
  * AUTHOR: Elizabeth Perreau and Dave Morton
  * DATE: 05-26-2014
@@ -53,7 +53,7 @@ $countSTH->bindValue(':bot_id', $bot_id, PDO::PARAM_INT);
 $countSTH->execute();
 $countRow = $countSTH->fetch();
 $countSTH->closeCursor();
-$row_count = number_format($countRow['count(id)']);
+$row_count = number_format($countRow['COUNT(id)']);
 
 $mainContent = str_replace('[row_count]', $row_count, $mainContent);
 $mainContent = str_replace('[bot_name]', $bot_name, $mainContent);