+++ gui/twitter/tweet.php
@@ -3,7 +3,7 @@
 /***************************************
  * http://www.program-o.com
  * PROGRAM O
- * Version: 2.6.3
+ * Version: 2.6.5
  * FILE: gui/plain/index.php
  * AUTHOR: Elizabeth Perreau and Dave Morton
  * DATE: MAY 17TH 2014
@@ -75,9 +75,9 @@ tweetthis($connection, $myReplies);
  *
  * Sends out tweets to the Twitter API, based on the chatbot's response
  *
- * @param (object) $connection
- * @param (array) $myReplies
- * @return (void)
+ * @param object $connection
+ * @param array $myReplies
+ * @return void
  */
 function tweetthis($connection, $myReplies)
 {
@@ -112,9 +112,9 @@ function tweetthis($connection, $myReplies)
  *
  * Collects the chatbot's response to the incoming tweet
  *
- * @param (string) $convo_id
- * @param (string) $usersay
- * @return (string) $botsay
+ * @param string $convo_id
+ * @param string $usersay
+ * @return string $botsay
  */
 function getReply($convo_id, $usersay)
 {
@@ -124,7 +124,7 @@ function getReply($convo_id, $usersay)
     $request_url = $chatbot_endpoint . "?say=" . urlencode($usersay) . "&convo_id=$convo_id&bot_id=$bot_id&format=xml";
     $conversation = @simplexml_load_file($request_url, "SimpleXmlElement", LIBXML_NOERROR + LIBXML_ERR_FATAL + LIBXML_ERR_NONE);
 
-    if ((@$conversation) && (count(@$conversation) > 0))
+    if ((@$conversation) && (@$conversation->count()) > 0)
     {
         $botsay = (string)$conversation->chat->line[0]->response;
         $botsay = str_replace("undefined", "...", $botsay);