+++ gui/twitter/twitteroauth/twitteroauth.php
@@ -76,6 +76,10 @@ class TwitterOAuth
 
     /**
      * construct TwitterOAuth object
+     * @param $consumer_key
+     * @param $consumer_secret
+     * @param null $oauth_token
+     * @param null $oauth_token_secret
      */
     function __construct($consumer_key, $consumer_secret, $oauth_token = NULL, $oauth_token_secret = NULL)
     {
@@ -92,7 +96,8 @@ class TwitterOAuth
     /**
      * Get a request_token from Twitter
      *
-     * @returns a key/value array containing oauth_token and oauth_token_secret
+     * @param $oauth_callback
+     * @return mixed A key/value array containing oauth_token and oauth_token_secret
      */
     function getRequestToken($oauth_callback)
     {
@@ -101,13 +106,16 @@ class TwitterOAuth
         $request = $this->oAuthRequest($this->requestTokenURL(), 'GET', $parameters);
         $token = OAuthUtil::parse_parameters($request);
         $this->token = new OAuthConsumer($token['oauth_token'], $token['oauth_token_secret']);
-        return $token;
+      /** @noinspection PhpIncompatibleReturnTypeInspection */
+      return $this->token;
     }
 
     /**
      * Get the authorize URL
      *
-     * @returns a string
+     * @param $token
+     * @param bool $sign_in_with_twitter
+     * @return string
      */
     function getAuthorizeURL($token, $sign_in_with_twitter = TRUE)
     {
@@ -115,9 +123,11 @@ class TwitterOAuth
             $token = $token['oauth_token'];
         }
         if (empty($sign_in_with_twitter)) {
-            return $this->authorizeURL() . "?oauth_token={$token}";
+          /** @noinspection PhpIncompatibleReturnTypeInspection */
+          return $this->authorizeURL() . "?oauth_token={$token}";
         } else {
-            return $this->authenticateURL() . "?oauth_token={$token}";
+          /** @noinspection PhpIncompatibleReturnTypeInspection */
+          return $this->authenticateURL() . "?oauth_token={$token}";
         }
     }
 
@@ -125,7 +135,8 @@ class TwitterOAuth
      * Exchange request token and secret for an access token and
      * secret, to sign API calls.
      *
-     * @returns array("oauth_token" => "the-access-token",
+     * @param $oauth_verifier
+     * @return array ("oauth_token" => "the-access-token",
      *                "oauth_token_secret" => "the-access-secret",
      *                "user_id" => "9436992",
      *                "screen_name" => "abraham")
@@ -143,7 +154,9 @@ class TwitterOAuth
     /**
      * One time exchange of username and password for access token and secret.
      *
-     * @returns array("oauth_token" => "the-access-token",
+     * @param $username
+     * @param $password
+     * @return array ("oauth_token" => "the-access-token",
      *                "oauth_token_secret" => "the-access-secret",
      *                "user_id" => "9436992",
      *                "screen_name" => "abraham",
@@ -163,6 +176,9 @@ class TwitterOAuth
 
     /**
      * GET wrapper for oAuthRequest.
+     * @param $url
+     * @param array $parameters
+     * @return mixed
      */
     function get($url, $parameters = array())
     {
@@ -175,6 +191,9 @@ class TwitterOAuth
 
     /**
      * POST wrapper for oAuthRequest.
+     * @param $url
+     * @param array $parameters
+     * @return mixed
      */
     function post($url, $parameters = array())
     {
@@ -187,6 +206,9 @@ class TwitterOAuth
 
     /**
      * DELETE wrapper for oAuthReqeust.
+     * @param $url
+     * @param array $parameters
+     * @return mixed
      */
     function delete($url, $parameters = array())
     {
@@ -199,6 +221,10 @@ class TwitterOAuth
 
     /**
      * Format and sign an OAuth / API request
+     * @param $url
+     * @param $method
+     * @param $parameters
+     * @return mixed
      */
     function oAuthRequest($url, $method, $parameters)
     {
@@ -218,7 +244,10 @@ class TwitterOAuth
     /**
      * Make an HTTP request
      *
-     * @return API results
+     * @param $url
+     * @param $method
+     * @param null $postfields
+     * @return mixed results
      */
     function http($url, $method, $postfields = NULL)
     {
@@ -259,6 +288,9 @@ class TwitterOAuth
 
     /**
      * Get the header info to store.
+     * @param $ch
+     * @param $header
+     * @return int
      */
     function getHeader($ch, $header)
     {